# docs/swagger.yaml
basePath: /
definitions:
  model.AssignPermissionsRequest:
    properties:
      permission_ids:
        items:
          type: string
        type: array
    required:
    - permission_ids
    type: object
...
  model.CreateRoleRequest:
    properties:
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        type: string
    required:
    - display_name
    - name
    type: object
...
  model.PageMetadata:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  model.RemovePermissionsRequest:
    properties:
      permission_ids:
        items:
          type: string
        type: array
    required:
    - permission_ids
    type: object
  model.SwaggerPageResponse:
    properties:
      data:
        items: {}
        type: array
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
  model.SwaggerWebResponse:
    properties:
      data: {}
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
...
  model.UpdateRoleRequest:
    properties:
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Golang Clean Architecture.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MKP Boarding Test API
  version: "1.0"
paths:
...
  /api/roles:
    get:
      consumes:
      - application/json
      description: Get list of roles with optional filtering
      parameters:
      - description: Filter by role name
        in: query
        name: name
        type: string
      - default: true
        description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            $ref: '#/definitions/model.SwaggerPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role with name and description
      parameters:
      - description: Create role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role created successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Roles
  /api/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Delete role by role ID
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get role details by role ID
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role details
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role information by role ID
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
  /api/roles/{roleId}/permissions:
    delete:
      consumes:
      - application/json
      description: Remove multiple permissions from a specific role
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Remove permissions request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RemovePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permissions removed successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Remove permissions from role
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Assign multiple permissions to a specific role
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Assign permissions request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permissions assigned successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Assign permissions to role
      tags:
      - Roles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"